Re-configured compilation and integrated uploading.  Added the Board and Serial port menus: these are currently shared across editor windows.
No longer using Library and LibraryManager.  Instead the compiler finds the .c and .cpp files in the library paths (gotten from the Sketch via the Preprocessor and the Base's library table).  This breaks libraries that have header files in sub-directories but use #include with no path names (as sub-directories of library directories are not on the include path).  It will also cause problems with .c or .cpp files with the same name in different library utility folders or in a sketch or the target (core).

Now deriving the Target correctly from the selected item in the board menu.



Former-commit-id: 14778514aec54c47df1b2839c0b382a52d838e18