Changing extension from .pde to .ino.
Renames all .pde files in a sketch to .ino upon opening.  Prompts the user the first time this happens; if the rename is cancelled, the sketch is not opened and the user will be prompted next time a sketch with a .pde file is opened.  Otherwise, renames files in all future opened sketches automatically (without prompting).  Still allows for opening of .pde files and includes them in the sketchbook and examples menus.

Still need to check the file association code.

http://code.google.com/p/arduino/issues/detail?id=13

Former-commit-id: e2213d4721ced1bd64242cd7274f028bcb95b996